Fix not fully cleared suffix in NextVersions
Extending |Version.String()| to cover the full length of a Version,
not only the first three columns, revealed this error in |NextVersions|
which yielded lists with Versions with not fully empty suffixes.

For example, 1.2.3-beta2 would have a successor "1.3.0[-common]2"
because after the "1.3"—which we want—only column 3 and 4 got set
to zero.

This didn't have any apparent effect because aforementioned |String|
was used, and incidentially covered only the parts that have been
set to zero ("cleared") properly.