update benchmarks for Go 1.12
Please note that every implementation is supplied the same version,
it's in different formats, the best suited ones without time lost
to any conversions. Those happen outside the benchmarks.
      This gives the contenders an advantage not encountered in
practice, because there []byte is the most common and would've needed
to be converted.

Further, to account for runs in single-digit nanoseconds, I've
changed aforementioned reference version representation to a longer
one. Which in turn will increase pressure on implementations with an
irregular core (branches, quirky data structures).