improve performance of atoui
Reduces branching from 4 to 3, and opcodes from 31 to 26.

Hot loop and preamble are one opcode shorter, but it's to subtle for
the benchmark on amd64 to be picked up.

Allow to use one more digit. For the full range we'd need to check
'carry' after the multiplication, which is currently not possible in
Go without resorting to Assembly completely.