use a sample distribution of Versions for benchmarks
In benchmarks, remove any bias inherent to fixed version strings
by using a sample distribution of versions encountered in the wild.

Benchmarks are henceforth ran on an AMD Zen 2 cpu (in ‘performance’
mode and disabled boost). My Haswell cpus are retired, and I anticipate
cloud providers to switch away in the mid- to long run.