sort: using radix sort variants
Versions in the wild are overwhelmingly low single digits, inviting
the specialized flavours of radix sort used here.

Any allocations but one to three (depending on the version of Go used)
are due to sort.Sort; sort.Slice would result in even more allocations
as it resorts to its own Swapper. Incidentially the latter is faster,
but I decided to keep the first implementation rather simple.