sort: implement twoFieldKey for amd64 using MMX
Decreases pressure on branch predictor for those branches that are
hard to predict anyway.

Using MMX allows for better performance with SMT as those sections
can be run concurrently to non-MMX/non-FP loads in an ideal way.
Especially on SMT4 chips.

Trades a performance hit of 1% to sorting on Silvermont and Zen1
to an increase of 1% on newer chips that'll be offset by a later
change allowing radix sort to descent to keyIndex 4 and below.