sort: offset negative fields and make them available to radix sort
Enables to run radix sort on all fields and skip any sort.Slice
or sort.Sort (were not the ramp-up costs for radix). In other words,
we could stay longer with the fast algorithm for larger data sets.

Due to the better alignment sorting became faster on Zen by 2–5%.
Involvement of MMX/FP units made it slower on Silvermont by -5%.
Performance decrease on Haswell and later is -1% or lower.