sort: use random versions as substitutes on parsing errors
Going with a constant whenever a library didn't parse a version correctly
results in a list with long strides that don't need to be sorted. Which
does not resemble what can be found in the wild.

I could pick a different entry from the list, but chances are that is
another which cannot be parsed (or does not conform to Semver indeed).
Therefore random suffix it is for now, by extension a random version.